Move SessionKey into a separate file, added comments
